[frontend] fix dependency cycle for deletion (#10985)
