+++ opencti-platform/opencti-front/src/private/components/workspaces/WorkspaceEditionQuery.tsx
@@ -1,0 +1,11 @@
+import { graphql } from 'react-relay';
+
+const WorkspaceEditionQuery = graphql`
+    query WorkspacePopoverContainerQuery($id: String!) {
+        workspace(id: $id) {
+            ...WorkspaceEditionContainer_workspace
+        }
+    }
+`;
+
+export default WorkspaceEditionQuery;
