+++ opencti-platform/opencti-front/src/private/components/workspaces/WorkspacePopoverDeletionMutation.tsx
@@ -1,0 +1,9 @@
+import { graphql } from 'react-relay';
+
+const WorkspacePopoverDeletionMutation = graphql`
+    mutation WorkspacePopoverDeletionMutation($id: ID!) {
+        workspaceDelete(id: $id)
+    }
+`;
+
+export default WorkspacePopoverDeletionMutation;
